+++ test/dummy/app/controllers/home_controller.rb
@@ -4,9 +4,10 @@ class HomeController < ApplicationController
   def index
     @url = params[:url].to_s
     @probe = params[:probe] == "1"
+    @approach = params[:approach] == "browser" ? "browser" : "download"
     return if @url.blank?
 
-    @page = RecordingStudio::WebReader.read(@url)
+    @page = RecordingStudio::WebReader.read(@url, strategy: @approach == "browser" ? :browser : :http)
     @page = RecordingStudio::WebReader.probe_images(@page) if @probe
     @paywall = @page.analyze(:paywall)
   rescue RecordingStudio::WebReader::Error => error